python-api/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
langchain==0.3.25
langchain-community==0.3.24
langchain-openai==0.3.21
pinecone-client==6.0.0
pydantic==2.11.5
pydantic-settings==2.9.1
python-dotenv==1.1.0
requests==2.32.3